All has been done
-Alot of components have been made. We made and edited the nav
component.
-We made and edited the home component
-and the same with alot of other stuff. The important things happen to
be in comments and are lessons mostly for html.

-An interesting thing was added though. An Authentication guard we put
so that only people who login can acess the web pages. Note that this
isn't securtiy. Someone can easily inject a user json in their local
storage and trick auth guard into letting them in

It will work easy if you follow along but we ran into some errors.
Like for example it will register a link but wont go through. That's
cause we didn't have and observable. Then the other error was cause we
din't have it to be there when ever we refreshed so we had to make
changes to the app component ts